% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donut.R
\name{donut_chart}
\alias{donut_chart}
\title{Creates a donut chart.}
\usage{
donut_chart(data, groupColumn, totalsColumn, centerLabel = "",
  centerColor = "#767676", centerLabelColor = "white",
  centerLabelSize = 4, startColor = "#00c896", endColor = "#678fdc",
  outerLabelColor = "black", outerLabelSize = 2.5,
  includePercentage = T, nudgeouterLabels = 0)
}
\arguments{
\item{data}{The data set.}

\item{groupColumn}{The column with the categorical labels.}

\item{totalsColumn}{The column with the categorical totals.}

\item{centerLabel}{The label displayed in the center of the donut.}

\item{centerColor}{The color of the donut's center.}

\item{centerLabelColor}{The font color of the donut's center label.}

\item{centerLabelSize}{The font size of the donut's center label.}

\item{startColor}{The starting color of the range of colors used for the slices.}

\item{endColor}{The ending color of the range of colors used for the slices.}

\item{outerLabelColor}{The font color of the outer labels.}

\item{outerLabelSize}{The font size of the outer labels.}

\item{includePercentage}{Whether to include percentages next to the slices.}

\item{nudgeouterLabels}{How far to nudge the outer labels away from the slices.}
}
\value{
A donut chart (a ggplot object).
}
\description{
Creates a donut chart.
}
\examples{
as_tibble(Titanic) \%>\% count(Class, wt=n) \%>\% rename(PassengersCount=nn) \%>\%
  donut_chart(Class, PassengersCount, "Class")
}
