% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donut.R
\name{donut_chart}
\alias{donut_chart}
\title{Creates a donut chart.}
\usage{
donut_chart(data, groupColumn, totalsColumn, centerLabel = "",
  centerColor = "#767676", centerLabelColor = "white",
  centerLabelSize = 24, startColor = "#00c896", endColor = "#678fdc",
  outerLabelColor = "black", outerLabelSize = 10,
  includePercentage = T, percentOnSeparateLine = T, ...)
}
\arguments{
\item{data}{The data set.}

\item{groupColumn}{The column with the categorical labels.}

\item{totalsColumn}{The column with the categorical totals.}

\item{centerLabel}{A character vector. The label displayed in the center of the donut.}

\item{centerColor}{A character vector. The color of the donut's center.}

\item{centerLabelColor}{A character vector. The font color of the donut's center label.}

\item{centerLabelSize}{An integer. The font (point) size of the donut's center label.}

\item{startColor}{The starting color of the range of colors used for the slices.}

\item{endColor}{A character vector. The ending color of the range of colors used for the slices.}

\item{outerLabelColor}{A character vector. The font color of the outer labels.}

\item{outerLabelSize}{An integer. The font (point) size of the outer labels.}

\item{includePercentage}{A logical. Whether to include percentages next to the slices.}

\item{percentOnSeparateLine}{A logical. Whether the percent label for each slice should be on a separate line (below the label). includePercentage must be TRUE to take effect.}

\item{...}{Additional arguments to pass to geom_text_repel (controls the slice labels).}
}
\description{
Creates a donut chart.
}
\examples{
as_tibble(Titanic) \%>\% count(Class, wt=n) \%>\% rename(PassengersCount=nn) \%>\%
  donut_chart(Class, PassengersCount, "Class")
}
